Update Docs and Response Codes
Updated the docs to reflect the current api

Changed the responses to action results so the correct response code is used

I'll probably have to do a re-write of the open api v3 spec because there seems to be a handful of errors that are hard to spot and things have changed since I wrote it. Also I need to figure out authorization before I do a re-write.